feat: Add QuestionController endpoints and error handling

+++ quiz-api/src/main/java/com/codersquiz/quiz_api/controllers/TopicController.java
@@ -1,30 +1,43 @@
 package com.codersquiz.quiz_api.controllers;
 
-import com.codersquiz.quiz_api.models.Starter;
+import com.codersquiz.quiz_api.exceptions.ResourceNotFoundException;
 import com.codersquiz.quiz_api.models.Topic;
 import com.codersquiz.quiz_api.repositories.TopicRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+
 import java.util.List;
 import java.util.Optional;
+
+
+
 
 @RestController
 @RequestMapping("/quiz-api/topics")
 public class TopicController {
+
+
     @Autowired
-//    private TopicService topicService;
     private TopicRepository topicRepository;
+
 
     @GetMapping("/all")
     public List<Topic> getAllTopics() {
         return topicRepository.findAll();
     }
 
+
     @GetMapping("/{topicId}")
     public Topic getTopic(@PathVariable("topicId") Long topicId) {
-        return topicRepository.findById(topicId).get();
+        Optional<Topic> optionalTopic = topicRepository.findById(topicId);
+        if (optionalTopic.isPresent()) {
+            return optionalTopic.get();
+        } else {
+            throw new ResourceNotFoundException("Topic not found with id " + topicId);
+        }
     }
+
 
     @PostMapping
     public Topic createTopic(@RequestBody String name) {
@@ -46,9 +59,15 @@ public class TopicController {
         }
     }
 
+
     @DeleteMapping("/{topicId}")
     public String deleteTopic(@PathVariable Long topicId) {
-        topicRepository.deleteById(topicId);
-        return "Topic deleted successfully";
+        Optional<Topic> optionalTopic = topicRepository.findById(topicId);
+        if (optionalTopic.isPresent()) {
+            topicRepository.deleteById(topicId);
+            return "Topic deleted successfully";
+        } else {
+            throw new ResourceNotFoundException("Topic not found with id " + topicId);
+        }
     }
 }
